refactor(cli): move shared visual/text helpers to tui_text.go (PRD #1009 M3)
